refactor(result): InternalResult into result module
