+++ src/serializer.rs
@@ -2,7 +2,8 @@ use serde::Serialize;
 
 use crate::{
     error::Error,
-    handler::{InternalResult, Json, StandardBodyType},
+    handler::{Json, StandardBodyType},
+    result::InternalResult,
 };
 
 pub trait BodySerializer {
